Better handling of unexpected results. (#2)
The script has a timeout of 600 seconds, after which point it printed the message "Some test runs did not complete." The timeout can be triggered in three different cases:
- The server returned `none` for the data field.
- The server returned an unrecognized result status.
- The tests are still running (the result status is "working").

I added print statements for the first two cases. Without them it's hard to know for sure why the script timed out. I also updated the timeout message to be more clear that there is an expected time window. And I moved the deadline check into the while loop because it's a little simpler. I think it makes it more obvious when the script exits due to timeout.

I tested these changes locally a bit and they seem ok.